Create a Dev Test UI Prefab Add a texture to the terrain
