+++ Assets/Digital Painting/Prefabs/DevTest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 482035e8dee8d2e4696be77242ccf719
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
