+++ Assets/Digital Painting/Prefabs/DevTest/Canvas.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c495122d1430c347ab9945831679e2b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
