+++ Assets/Digital Painting/Scenes/Demo - Standard Assets Only.unity
@@ -304,8 +304,8 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 403725151}
-  m_LocalRotation: {x: -0.011763349, y: -0.96347815, z: 0.26406324, w: -0.04292057}
-  m_LocalPosition: {x: 3.4856415, y: -71.041504, z: 33.735977}
+  m_LocalRotation: {x: -0.11937078, y: 0.8604847, z: -0.23583555, w: -0.43554372}
+  m_LocalPosition: {x: 11.80658, y: -71.04148, z: 32.290543}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 1500001151}
@@ -515,8 +515,8 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 967244701}
-  m_LocalRotation: {x: 0.0167821, y: -0.8862123, z: 0.46185413, w: 0.03220214}
-  m_LocalPosition: {x: 8.184021, y: -67.94317, z: 31.060013}
+  m_LocalRotation: {x: 0.20473708, y: 0.7950323, z: -0.41433516, w: 0.3928521}
+  m_LocalPosition: {x: 2.819336, y: -67.94313, z: 30.129196}
   m_LocalScale: {x: 1.0000002, y: 1, z: 1.0000007}
   m_Children:
   - {fileID: 2102782024}
@@ -654,8 +654,8 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1372618687}
-  m_LocalRotation: {x: 0.018179886, y: -0.032458354, z: 0.00059057336, w: 0.9993076}
-  m_LocalPosition: {x: 50, y: 3.534, z: 40}
+  m_LocalRotation: {x: 0.0069378046, y: 0.9242511, z: -0.016814396, w: 0.3813517}
+  m_LocalPosition: {x: 248.85335, y: 42.36, z: 281.8331}
   m_LocalScale: {x: 1, y: 0.1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
