--- Assets/Digital Painting/Scripts/Environment/DayNightCycle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b734a48009a61b47aa78c2cb2073e12
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
